Extract test cases for assignment statements into separate file. Add parsing function to custom helper module.
